🚚 rename [#26] 파일 경로 이동: `detail-card` -> `card`
