+++ src/features/card/apis/index.ts
@@ -1,6 +1,6 @@
 import { type Card } from "@/types/card";
-import CardMock from "./card-mock.json";
+import cardMock from "./card-mock.json";
 
 export async function getCard(): Promise<Card> {
-  return CardMock;
+  return cardMock;
 }
