+++ src/features/card/components/card-detail-modal.module.css
