+++ src/pages/test-components/index.tsx
@@ -21,13 +21,13 @@ import SheetSectionGroup from "@/components/sheet/sheet-section-group";
 import Typography from "@/components/typography";
 import { CommonSize } from "@/constants/common/common-size";
 import { ProfileRandomColor } from "@/constants/profile-random-color";
+import { getCard } from "@/features/card/apis";
+import CardDetailModal from "@/features/card/components/card-detail-modal";
 import Dropdown, { DropdownOption } from "@/features/card/components/dropdown";
 import ImageInput from "@/features/card/components/image-input";
 import { Direction, Menu, MenuItem } from "@/features/card/components/menu";
 import TagInput from "@/features/card/components/tag-input";
 import ColumnEditSheet from "@/features/column/components/column-edit-sheet";
-import { getCard } from "@/features/detail-card/apis/card";
-import CardDetailModal from "@/features/detail-card/components/card-detail-modal";
 import { useAlert } from "@/hooks/use-alert";
 import { useDialog } from "@/hooks/use-dialog";
 import { useModal } from "@/hooks/use-modal";
